Add files via upload
This is the first part of flight delay prediction i.e. for the month of January. To check the second part i.e. for the month of February, please have a look at this dataset: February Flight Delay Prediction

This data is collected from the Bureau of Transportation Statistics, Govt. of the USA. This data is open-sourced under U.S. Govt. Works. This dataset contains all the flights in the month of January 2019 and January 2020. There are more than 400,000 flights in the month of January itself throughout the United States. The features were manually chosen to do a primary time series analysis. There are several other features available on their website.

This data could well be used to predict the flight delay at the destination airport specifically for the month of January in upcoming years as the data is for January only.